+++ presentation/src/main/java/com/chobo/presentation/view/event/component/Events.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.unit.dp
 import com.chobo.domain.model.event.response.GetEventListResponseModel
 import com.chobo.presentation.R
 import com.chobo.presentation.view.component.icon.ChevronRightIcon
-import com.chobo.presentation.view.component.multipleEventsCutterManager.clickableSingle
+import com.chobo.presentation.view.component.modifier.multipleEventsCutterManager.clickableSingle
 import com.chobo.presentation.view.theme.MindWayAndroidTheme
 import com.chobo.presentation.viewModel.util.formatServerDate
