+++ presentation/src/main/java/com/chobo/presentation/view/main/screen/ViewDetailScreen.kt
@@ -31,7 +31,7 @@ import com.chobo.presentation.view.component.bottom_sheet.MindWayBottomSheet
 import com.chobo.presentation.view.component.bottom_sheet.MindWayBottomSheetDialog
 import com.chobo.presentation.view.component.icon.ChevronLeftIcon
 import com.chobo.presentation.view.component.icon.OptionIcon
-import com.chobo.presentation.view.component.multipleEventsCutterManager.clickableSingle
+import com.chobo.presentation.view.component.modifier.multipleEventsCutterManager.clickableSingle
 import com.chobo.presentation.view.component.topBar.MindWayTopAppBar
 import com.chobo.presentation.view.main.component.ViewDetailPopUp
 import com.chobo.presentation.view.main.component.ViewDetailTextCard
