+++ presentation/src/main/java/com/chobo/presentation/view/my/component/MyBookListItem.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.chobo.presentation.view.component.icon.EditIcon
 import com.chobo.presentation.view.component.icon.TrashCanIcon
-import com.chobo.presentation.view.component.multipleEventsCutterManager.clickableSingle
+import com.chobo.presentation.view.component.modifier.multipleEventsCutterManager.clickableSingle
 import com.chobo.presentation.view.theme.MindWayAndroidTheme
 @Stable
 @Composable
